# BackEnd/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "XMeme API",
        "description": "API for a Meme Stream APP",
        "version": "1.0.0",
        "contact": {
            "email": "arijitroyb@gmail.com"
        },
        "license": {
            "name": "AGPLv3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        }
    },
    "host": "localhost:8081",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/memes/": {
            "get": {
                "tags": [
                    "XMeme API"
                ],
                "description": "Get Latest 100 Memes",
                "operationId": "MemeController.GetAllMemes",
                "responses": {
                    "200": {
                        "description": "\"Retrieved latest 100 memes\"",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Meme"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },

                    "404": {
                        "description": "Not Found!!"
                    }
                }
            },
            "post": {
                "tags": [
                    "XMeme API"
                ],
                "description": "Post New Meme",
                "operationId": "MemeController.CreateMeme",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Meme Request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MemeForPost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\"Meme created successfully\"",
                        "schema": {
                            "$ref": "#/definitions/models.IResponse"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "Not Found!!"
                    }
                }
            }
        },
        "/memes/{id}": {
            "get": {
                "tags": [
                    "XMeme API"
                ],
                "description": "Get Meme by Id",
                "operationId": "MemeController.GetMeme",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "the id you want to get",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\"Retrieved Meme from given id successfully\"",
                        "schema": {
                            "$ref": "#/definitions/models.MemeForPost"
                        }
                    },
                    "404": {
                        "description": "Meme not found. Provide valid meme-id"
                    },
                    "406": {
                        "description": ":id is empty"
                    }
                }
            },
            "patch": {
                "tags": [
                    "XMeme API"
                ],
                "description": "Update Meme by Id",
                "operationId": "MemeController.UpdateMeme",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "Id of meme you want to update",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }, 
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Update meme",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MemeForPatch"
                        }
                    }
            
                ],
                "responses": {
                    "200": {
                        "description": "\"Success\""
                    },
                    "404": {
                        "description": "Not Found!! Send valid Id"
                    },
                    "406": {
                        "description": ":id is empty"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.IResponse": {
            "title": "IResponse",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "models.Meme": {
            "title": "Meme",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }, "url": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                }
            }
        },
        "models.MemeForPost": {
            "title": "Meme",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }, "url": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                }
            }
        },
        "models.MemeForPatch": {
            "title": "Meme",
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                }
            }
        }
    }
}